refactored Scoreboard entry to own class
